refactored where render logic is, making new branch for further changes
