+++ client/src/views/map.js
@@ -1,4 +1,5 @@
 var Game = require('./game.js');
+var InventoryUI = require('./inventoryUI.js');
 var Player = require('../models/player_model.js');
 var Bar = require('../models/bar_model.js');
 
@@ -12,6 +13,7 @@ var Map = function () {
     cashDrawer: 1000
   });
   this.game = new Game(this.player, this.bar);
+  this.inventoryUI = new InventoryUI(this.player, this.bar);
 
   context = getCanvasContext();
